requirements.txt
streamlit
transformers
torch
torchvision
pillow
openai